@@ -1,6 +1,5 @@
 ﻿// (c) Johannes Wolfgruber, 2020
 
-using NAudio.Wave;
 using SkiaSharp;
 using SkiaSharp.Views.Desktop;
 using SkiaSharp.Views.WPF;
@@ -15,79 +14,80 @@ namespace SystemAudioRecordingSoftware.UI.Views
 {
     public class Waveform
     {
-        private readonly List<double> _snipTimestamps = new List<double>();
-        private float[]? _audioData;
-        private int _zoomFactor = 1000;
-        private double _lengthInSeconds;
-        private double _currentWidth;
-        private double _currentTimestamp;
+        private const double MaxWidth = 2500;
+        private readonly List<float> _audioData = new();
+        private readonly List<TimeSpan> _snipTimestamps = new();
+        private double _currentAudioWidth;
+        private TimeSpan _currentTimestamp;
+        private TimeSpan _lengthInSeconds;
+        private int _numberOfSamples;
+        private int _zoomFactor = 10;
 
-        public float StrokeWidth { get; set; } = 1f;
+        public float StrokeWidth { get; set; } = .5f;
         public SKColor Color { get; set; } = SKColors.White;
 
-        public void SetWaveFile(string filePath)
+        public void AddAudioData(IEnumerable<float> buffer, int totalNumberOfSamples, int sampleRate)
         {
-            if (string.IsNullOrEmpty(filePath))
-            {
-                return;
-            }
-
-            using var reader = new AudioFileReader(filePath);
-            float[] buffer = new float[reader.Length / 2];
-            reader.Read(buffer, 0, buffer.Length);
-            _audioData = new float[buffer.Length / 2];
-            Array.Copy(buffer, 0, _audioData, 0, buffer.Length / 2);
-            _lengthInSeconds = (double)_audioData.Length / reader.WaveFormat.SampleRate;
+            _audioData.AddRange(buffer);
+            _numberOfSamples += totalNumberOfSamples;
+            _lengthInSeconds = TimeSpan.FromSeconds((double)_numberOfSamples / sampleRate);
         }
 
-        public void SetClickPosition(Point clickPosition, TextBlock textBlock)
+        public void AddLiveSnipPosition()
         {
-            _currentTimestamp = ToTimestamp(clickPosition.X);
-            textBlock.Text = $"X: {clickPosition.X}, Y: {clickPosition.Y}, timestamp: {_currentTimestamp}s";
-        }
-
-        private double ToTimestamp(double x)
-        {
-            return (_lengthInSeconds * x) / _currentWidth;
-        }
-
-        private double ToX(double timestamp)
-        {
-            return (_currentWidth * timestamp) / _lengthInSeconds;
-        }
-
-        public void ZoomOut()
-        {
-            _zoomFactor += 100;
-        }
-
-        public void ZoomIn()
-        {
-            if (_zoomFactor != 100)
-            {
-                _zoomFactor -= 100;
-            }
-        }
-
-        public void SetWidth(double parentWidth, SKElement element)
-        {
-            if (_audioData == null)
-            {
-                return;
-            }
-
-            var newWidth = _audioData.Length / _zoomFactor;
-            if (newWidth >= parentWidth)
-            {
-                element.Width = newWidth;
-            }
-
-            _currentWidth = element.ActualWidth;
+            _snipTimestamps.Add(_lengthInSeconds);
         }
 
         public void AddSnipPosition()
         {
             _snipTimestamps.Add(_currentTimestamp);
+        }
+
+        public void DrawOnCanvas(SKCanvas canvas)
+        {
+            if (_audioData.Count <= 0)
+            {
+                return;
+            }
+
+            canvas.Clear(SKColors.Transparent);
+
+            var dimensions = canvas.DeviceClipBounds;
+            var midpoint = dimensions.Height / 2;
+            var width = dimensions.Width;
+            var hopSize = _zoomFactor;
+
+            var paint = CreatePaint(Color);
+
+            for (var i = 0; i < width; i++)
+            {
+                var amplitudeValue = (i + 1) * hopSize >= _audioData.Count
+                    ? 0
+                    : GetAbsoluteMaxValue(_audioData.ToArray(), i * hopSize, (i + 1) * hopSize) * midpoint;
+                canvas.DrawLine(
+                    new SKPoint(i, midpoint + Math.Abs(amplitudeValue)),
+                    new SKPoint(i, midpoint - Math.Abs(amplitudeValue)),
+                    paint);
+            }
+
+            canvas.DrawLine(
+                new SKPoint((float)(ToX(_currentTimestamp)), 0),
+                new SKPoint((float)(ToX(_currentTimestamp)), canvas.DeviceClipBounds.Height),
+                paint);
+
+            canvas.DrawLine(
+                new SKPoint((float)(ToX(_lengthInSeconds)), 0),
+                new SKPoint((float)(ToX(_lengthInSeconds)), canvas.DeviceClipBounds.Height),
+                paint);
+
+            var redPaint = CreatePaint(SKColors.Red);
+            foreach (var timestamp in _snipTimestamps)
+            {
+                canvas.DrawLine(
+                    new SKPoint((float)(ToX(timestamp)), 0),
+                    new SKPoint((float)(ToX(timestamp)), canvas.DeviceClipBounds.Height),
+                    redPaint);
+            }
         }
 
         public void RemoveLastSnipPosition()
@@ -100,68 +100,54 @@ namespace SystemAudioRecordingSoftware.UI.Views
             _snipTimestamps.RemoveAt(_snipTimestamps.Count - 1);
         }
 
-        public void DrawOnCanvas(SKCanvas canvas)
+        public void Reset()
         {
-            if (_audioData == null)
+            _audioData.Clear();
+            _numberOfSamples = 0;
+            _lengthInSeconds = TimeSpan.Zero;
+            _currentTimestamp = TimeSpan.Zero;
+            _snipTimestamps.Clear();
+            _zoomFactor = 10;
+        }
+
+        public void SetClickPosition(Point clickPosition, TextBlock textBlock)
+        {
+            _currentTimestamp = ToTimestamp(clickPosition.X);
+            textBlock.Text =
+                $"X: {clickPosition.X}, Y: {clickPosition.Y}, timestamp: {_currentTimestamp.TotalSeconds}s";
+        }
+
+        public void SetWidth(double parentWidth, SKElement element)
+        {
+            if (_audioData.Count <= 0)
             {
                 return;
             }
 
-            canvas.Clear(SKColors.Transparent);
-
-            var dimensions = canvas.DeviceClipBounds;
-            var midpoint = dimensions.Height / 2;
-            var width = dimensions.Width;
-            var hopSize = (int)Math.Round((double)_audioData.Length / width);
-
-            var paint = CreatePaint(Color);
-            var path = new SKPath();
-            path.MoveTo(new SKPoint(0, midpoint));
-
-            for (var i = 0; i < width; i++)
+            _currentAudioWidth = (double)_audioData.Count / _zoomFactor;
+            if (_currentAudioWidth >= parentWidth && _currentAudioWidth <= MaxWidth)
             {
-                var from = i * hopSize;
-                var to = (i + 1) * hopSize;
-                var amplitudeValue = to >= _audioData.Length ? 0 : GetAbsoluteMaxValue(_audioData, from, to);
-
-                var multiplier = i % 2 == 0 ? 1 : -1;
-
-                var controlPoint = new SKPoint
-                {
-                    X = (float)(i - 0.5),
-                    Y = midpoint + (amplitudeValue * 100 * multiplier)
-                };
-
-                var amplitudePoint = new SKPoint
-                {
-                    X = i,
-                    Y = midpoint
-                };
-
-                path.QuadTo(controlPoint, amplitudePoint);
+                element.Width = _currentAudioWidth;
             }
 
-            canvas.DrawPath(path, paint);
-            canvas.DrawLine(
-                new SKPoint((float)(ToX(_currentTimestamp)), 0), 
-                new SKPoint((float)(ToX(_currentTimestamp)), canvas.DeviceClipBounds.Height), 
-                paint);
-
-            var redPaint = CreatePaint(SKColors.Red);
-            foreach (var timestamp in _snipTimestamps)
+            if (_currentAudioWidth >= MaxWidth)
             {
-                canvas.DrawLine(
-                    new SKPoint((float)(ToX(timestamp)), 0), 
-                    new SKPoint((float)(ToX(timestamp)), canvas.DeviceClipBounds.Height), 
-                    redPaint);
+                _zoomFactor++;
+                SetWidth(parentWidth, element);
             }
         }
 
-        private static float GetAbsoluteMaxValue(float[] audioData, int from, int to)
+        public void ZoomIn()
         {
-            var temp = new float[to - from];
-            Array.Copy(audioData, from, temp, 0, to - from);
-            return temp.Select(Math.Abs).Max();
+            if (_zoomFactor != 2)
+            {
+                _zoomFactor -= 1;
+            }
+        }
+
+        public void ZoomOut()
+        {
+            _zoomFactor += 1;
         }
 
         private SKPaint CreatePaint(SKColor color)
@@ -173,57 +159,81 @@ namespace SystemAudioRecordingSoftware.UI.Views
                 Style = SKPaintStyle.Stroke,
                 StrokeWidth = StrokeWidth
             };
+
+        private static float GetAbsoluteMaxValue(float[] audioData, int from, int to)
+        {
+            var temp = new float[to - from];
+            Array.Copy(audioData, from, temp, 0, to - from);
+            return temp.Select(Math.Abs).Max();
+        }
+
+        private TimeSpan ToTimestamp(double x)
+        {
+            return TimeSpan.FromSeconds((_lengthInSeconds.TotalSeconds * x) / ((double)_audioData.Count / _zoomFactor));
+        }
+
+        private double ToX(TimeSpan timestamp)
+        {
+            return (((double)_audioData.Count / _zoomFactor) * timestamp.TotalSeconds) / _lengthInSeconds.TotalSeconds;
+        }
     }
 
     public partial class WaveformRenderer
     {
-        public static readonly DependencyProperty FilePathProperty =
-            DependencyProperty.Register("FilePath", typeof(string), typeof(WaveformRenderer), new PropertyMetadata(default));
-
         private readonly Waveform _waveform = new Waveform();
+        private bool _autoScroll = true;
 
         public WaveformRenderer()
         {
             InitializeComponent();
         }
 
-        public string FilePath
+        public void AddAudioData(IEnumerable<float> buffer, int totalNumberOfSamples, int sampleRate)
         {
-            get { return (string)GetValue(FilePathProperty); }
-            set
-            {
-                SetValue(FilePathProperty, value);
-                OnFilePathChanged(value);
-            }
-        }
-
-        private void OnFilePathChanged(string filePath)
-        {
-            _waveform.SetWaveFile(filePath);
+            _waveform.AddAudioData(buffer, totalNumberOfSamples, sampleRate);
             SkiaElement.InvalidateVisual();
         }
 
-        private void SkiaElement_PaintSurface(object sender, SKPaintSurfaceEventArgs e)
+        public void AddLiveSnip()
         {
-            _waveform.SetWidth(ActualWidth, SkiaElement);
-            _waveform.DrawOnCanvas(e.Surface.Canvas);
+            _waveform.AddLiveSnipPosition();
+            SkiaElement.InvalidateVisual();
+        }
+
+        public void Reset()
+        {
+            SkiaElement.Width = ActualWidth;
+            _waveform.Reset();
+            SkiaElement.InvalidateVisual();
+        }
+
+        private void RemoveSnipButton_Click(object sender, RoutedEventArgs e)
+        {
+            _waveform.RemoveLastSnipPosition();
+            SkiaElement.InvalidateVisual();
+        }
+
+        private void ScrollViewer_OnScrollChanged(object sender, ScrollChangedEventArgs e)
+        {
+            if (!(sender is ScrollViewer scrollViewer))
+            {
+                return;
+            }
+
+            if (e.ExtentWidthChange == 0)
+            {
+                _autoScroll = scrollViewer.HorizontalOffset == scrollViewer.ScrollableWidth;
+            }
+
+            if (_autoScroll && e.ExtentWidthChange != 0)
+            {
+                scrollViewer.ScrollToHorizontalOffset(scrollViewer.ExtentWidth);
+            }
         }
 
         private void SkiaElement_MouseDown(object sender, MouseButtonEventArgs e)
         {
             _waveform.SetClickPosition(e.GetPosition(SkiaElement), LastClickPosition);
-            SkiaElement.InvalidateVisual();
-        }
-
-        private void ZoomOutButton_Click(object sender, RoutedEventArgs e)
-        {
-            _waveform.ZoomOut();
-            SkiaElement.InvalidateVisual();
-        }
-
-        private void ZoomInButton_Click(object sender, RoutedEventArgs e)
-        {
-            _waveform.ZoomIn();
             SkiaElement.InvalidateVisual();
         }
 
@@ -235,18 +245,6 @@ namespace SystemAudioRecordingSoftware.UI.Views
             }
 
             _waveform.SetClickPosition(e.GetPosition(SkiaElement), LastClickPosition);
-            SkiaElement.InvalidateVisual();
-        }
-
-        private void SnipButton_Click(object sender, RoutedEventArgs e)
-        {
-            _waveform.AddSnipPosition();
-            SkiaElement.InvalidateVisual();
-        }
-
-        private void RemoveSnipButton_Click(object sender, RoutedEventArgs e)
-        {
-            _waveform.RemoveLastSnipPosition();
             SkiaElement.InvalidateVisual();
         }
 
@@ -262,6 +260,30 @@ namespace SystemAudioRecordingSoftware.UI.Views
                 _waveform.ZoomOut();
                 SkiaElement.InvalidateVisual();
             }
+        }
+
+        private void SkiaElement_PaintSurface(object sender, SKPaintSurfaceEventArgs e)
+        {
+            _waveform.SetWidth(ActualWidth, SkiaElement);
+            _waveform.DrawOnCanvas(e.Surface.Canvas);
+        }
+
+        private void SnipButton_Click(object sender, RoutedEventArgs e)
+        {
+            _waveform.AddSnipPosition();
+            SkiaElement.InvalidateVisual();
+        }
+
+        private void ZoomInButton_Click(object sender, RoutedEventArgs e)
+        {
+            _waveform.ZoomIn();
+            SkiaElement.InvalidateVisual();
+        }
+
+        private void ZoomOutButton_Click(object sender, RoutedEventArgs e)
+        {
+            _waveform.ZoomOut();
+            SkiaElement.InvalidateVisual();
         }
     }
 }
