@@ -1,8 +1,8 @@
 ﻿// (c) Johannes Wolfgruber, 2020
+
 using NAudio.CoreAudioApi;
 using NAudio.Wave;
 using System;
-using SystemAudioRecordingSoftware.Core.Audio;
 using SystemAudioRecordingSoftware.Core.Model;
 
 namespace SystemAudioRecordingSoftware.Core.AudioEngine
@@ -12,11 +12,13 @@ namespace SystemAudioRecordingSoftware.Core.AudioEngine
         bool IsRecording { get; }
         IObservable<CaptureState> CaptureStateChanged { get; }
         IObservable<StoppedEventArgs> RecordingStopped { get; }
-        IObservable<MinMaxValuesEventArgs> SampleAvailable { get; }
+        IObservable<AudioDataDto> AudioDataAvailable { get; }
         IObservable<Recording> NewRecordingCreated { get; }
 
         void StartRecording();
 
         void StopRecording();
+        
+        void SnipRecording();
     }
 }
