@@ -1,43 +1,55 @@
 ﻿// (c) Johannes Wolfgruber, 2020
 
-using System;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Reactive;
-using SystemAudioRecordingSoftware.Core.Audio;
-using SystemAudioRecordingSoftware.Core.AudioEngine;
+using DynamicData;
+using NAudio.CoreAudioApi;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using Splat;
+using System;
+using System.Collections.ObjectModel;
+using System.Reactive;
+using System.Reactive.Linq;
+using SystemAudioRecordingSoftware.Core.AudioEngine;
+using SystemAudioRecordingSoftware.Core.Model;
 
 namespace SystemAudioRecordingSoftware.UI.ViewModels
 {
     public class MainWindowViewModel : ReactiveObject
     {
         private readonly IAudioEngineService _engineService;
-        private readonly PolygonWaveFormViewModel _visualization = new PolygonWaveFormViewModel();
+        private readonly ReadOnlyObservableCollection<RecordingViewModel> _recordings;
+        private readonly WaveformRendererViewModel _waveformRenderer = new WaveformRendererViewModel();
 
         public MainWindowViewModel(IAudioEngineService? engineService = null)
         {
             _engineService = engineService ?? Locator.Current.GetService<IAudioEngineService>();
 
-            _engineService.CaptureStateChanged
-                .Subscribe(_ => IsRecording = _engineService.IsRecording);
+            _engineService.CaptureStateChanged.Subscribe(OnCaptureStateChanged);
+            _engineService.AudioDataAvailable.Subscribe(OnAudioDataAvailable);
+            _engineService.RecordingsChanged()
+                .Transform(r => new RecordingViewModel(r))
+                .ObserveOn(RxApp.MainThreadScheduler)
+                .Bind(out _recordings)
+                .Subscribe();
 
-            _engineService.SampleAvailable.Subscribe(OnSampleAvailable);
+            var canStopOrSnip = this.WhenAnyValue(x => x.IsRecording);
+            var canRecordOrBurn = this.WhenAnyValue(x => x.IsRecording, isRecording => !isRecording);
+            var canDelete = this.WhenAnyValue(
+                x => x.SelectedRecording, 
+                x => x.SelectedTrack, 
+                (recording, track) => recording != null || track != null);
 
-            _engineService.RecordingsChanged
-                .Subscribe(x => Recordings = new ObservableCollection<RecordingViewModel>(
-                    x.Select(r => new RecordingViewModel(r))));
+            RecordCommand = ReactiveCommand.Create(OnRecord, canRecordOrBurn);
+            StopCommand = ReactiveCommand.Create(OnStop, canStopOrSnip);
+            SnipCommand = ReactiveCommand.Create(OnSnip, canStopOrSnip);
+            BurnCommand = ReactiveCommand.Create(OnBurn, canRecordOrBurn);
+            DeleteCommand = ReactiveCommand.Create(OnDelete, canDelete);
 
-            var canStop = this.WhenAnyValue(x => x.IsRecording);
-
-            RecordCommand = ReactiveCommand.Create(OnRecord);
-            StopCommand = ReactiveCommand.Create(OnStop, canStop);
-            Recordings = new ObservableCollection<RecordingViewModel>();
-
-            this.WhenAnyValue(x => x.SelectedRecording).Subscribe(x => FilePath = x?.FilePath);
+            this.WhenAnyValue(x => x.SelectedRecording)
+                .Subscribe(OnSelectedRecordingChanged);
         }
+
+        public ReadOnlyObservableCollection<RecordingViewModel> Recordings => _recordings;
 
         [Reactive] public bool IsRecording { get; set; }
 
@@ -45,24 +57,77 @@ namespace SystemAudioRecordingSoftware.UI.ViewModels
 
         public ReactiveCommand<Unit, Unit> StopCommand { get; }
 
-        [Reactive] public string Title { get; set; } = "System Audio Recording Software";
+        public ReactiveCommand<Unit, Unit> SnipCommand { get; }
 
-        [Reactive] public ObservableCollection<RecordingViewModel> Recordings { get; set; }
+        public ReactiveCommand<Unit, Unit> BurnCommand { get; }
+
+        public ReactiveCommand<Unit, Unit> DeleteCommand { get; }
+
+        [Reactive] public string Title { get; set; } = "System Audio Recording Software";
 
         [Reactive] public RecordingViewModel? SelectedRecording { get; set; }
 
-        public object Visualization => _visualization.Content;
+        [Reactive] public TrackViewModel? SelectedTrack { get; set; }
+
+        public object WaveformRenderer => _waveformRenderer.Content;
 
         [Reactive] public string? FilePath { get; set; }
+
+        public void OnSelectedItemChanged(object selectedItem)
+        {
+            SelectedRecording = selectedItem as RecordingViewModel;
+            SelectedTrack = selectedItem as TrackViewModel;
+        }
+
+        private void OnAudioDataAvailable(AudioDataDto args)
+        {
+            _waveformRenderer.AddAudioData(args.Buffer, args.TotalNumberOfSingleChannelSamples, args.SampleRate);
+        }
+
+        private void OnBurn()
+        {
+            throw new NotImplementedException();
+        }
+
+        private void OnCaptureStateChanged(CaptureState _)
+        {
+            IsRecording = _engineService.IsRecording;
+        }
+
+        private void OnDelete()
+        {
+            if (SelectedRecording != null)
+            {
+                _engineService.RemoveRecording(SelectedRecording.Id);
+            }
+        }
 
         private void OnRecord()
         {
             _engineService.Record();
+            _waveformRenderer.Reset();
         }
 
-        private void OnSampleAvailable(MinMaxValuesEventArgs args)
+        private void OnSelectedRecordingChanged(RecordingViewModel? vm)
         {
-            _visualization.OnSampleAvailable(args.MinValue, args.MaxValue);
+            FilePath = vm?.FilePath;
+            
+            if (FilePath is null)
+            {
+                return;
+            }
+            
+            _waveformRenderer.Reset();
+
+            var data = _engineService.GetAudioDisplayData(FilePath);
+
+            _waveformRenderer.AddAudioData(data.Buffer, data.TotalNumberOfSingleChannelSamples, data.SampleRate);
+        }
+
+        private void OnSnip()
+        {
+            _engineService.SnipRecording();
+            _waveformRenderer.AddSnip();
         }
 
         private void OnStop()
